Assignment 11

+++ java-assignments/Assignment_09/src/main/java/com/example/Assignment_09/service/PersonServiceImpl.java
@@ -53,8 +53,11 @@ public class PersonServiceImpl implements PersonService {
 
     @Override
     public boolean deletePerson(int id) {
-        Person person = getPersonById(id);
         try {
+            Person person = getPersonById(id);
+            if(person==null){
+                throw new Exception();
+            }
             personRepository.delete(person);
         } catch (Exception exception) {
             return false;
